requirements.txt
langchain-community==0.0.24
langchain-core>=0.1.26,<0.2.0
python-dotenv==1.0.0
pypdf==4.0.1
pandas==2.1.4
python-docx==1.1.0
chromadb==0.4.22
streamlit==1.32.0
transformers>=4.38.1
torch>=2.2.0
sentence-transformers>=2.5.1 